tratamento de exceções

+++ dscommerce/src/main/java/com_brandao/dscommerce/services/ProductService.java
@@ -1,14 +1,19 @@
 package com_brandao.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com_brandao.dscommerce.dtos.ProductDTO;
 import com_brandao.dscommerce.entities.Product;
 import com_brandao.dscommerce.repositories.ProductRepository;
+import com_brandao.dscommerce.services.exceptions.DatabaseException;
+import com_brandao.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -19,8 +24,10 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
 
-        Product product = repository.findById(id).get();        
-        return new ProductDTO(product);        
+        Product product = repository.findById(id).orElseThrow(
+                        () -> new ResourceNotFoundException("not found resource"));        
+        
+                        return new ProductDTO(product);        
     }
 
     @Transactional(readOnly = true)
@@ -46,21 +53,32 @@ public class ProductService {
     @Transactional
     public ProductDTO update(ProductDTO dto, Long id){
         
-        Product entity = repository.getReferenceById(id);
+        try
+        {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(entity, dto);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+
+        } catch(EntityNotFoundException e){
+            
+            throw new ResourceNotFoundException("resource not found");
+        }
         
-        copyDtoToEntity(entity, dto);
-
-        entity = repository.save(entity);
-
-        return new ProductDTO(entity);
 
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
 
-        repository.deleteById(id);
-
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("resource not found");
+        }
+        try{
+            repository.deleteById(id);
+        }catch(DataIntegrityViolationException e){
+            throw new DatabaseException("referencial integrity failure");
+        }
     }
